@@ -1,5 +1,5 @@
 {
-	"name": "ckbeamsoftware",
+	"name": "test-app",
 	"version": "0.1.0",
 	"private": true,
 	"scripts": {
@@ -9,53 +9,20 @@
 		"lint": "next lint"
 	},
 	"dependencies": {
-		"@heroicons/react": "^2.0.18",
-		"@radix-ui/react-dropdown-menu": "^2.0.6",
-		"@radix-ui/react-label": "^2.0.2",
-		"@radix-ui/react-navigation-menu": "^1.1.4",
-		"@radix-ui/react-slot": "^1.0.2",
-		"@radix-ui/react-switch": "^1.0.3",
-		"@radix-ui/react-toggle": "^1.0.3",
-		"class-variance-authority": "^0.7.0",
-		"clsx": "^2.0.0",
-		"gray-matter": "^4.0.3",
-		"html-react-parser": "^5.0.6",
-		"lucide-react": "^0.292.0",
-		"marked": "^10.0.0",
-		"next": "13.5.4",
-		"next-themes": "^0.2.1",
+		"next": "14.2.24",
 		"react": "^18",
 		"react-dom": "^18",
-		"react-loqate": "^1.3.0",
-		"tailwind-merge": "^2.0.0",
-		"tailwindcss-animate": "^1.0.7",
-		"usehooks-ts": "^2.9.1"
+		"react-slick": "^0.30.3",
+		"slick": "^1.12.2",
+		"slick-carousel": "^1.8.1"
 	},
 	"devDependencies": {
-		"@testing-library/jest-dom": "^5.16.5",
-		"@testing-library/react": "^14.0.0",
-		"@testing-library/user-event": "^14.4.3",
 		"@types/node": "^20",
-		"@types/react": "^18.2.35",
+		"@types/react": "^18",
 		"@types/react-dom": "^18",
-		"@types/testing-library__react": "^10.2.0",
-		"@types/whatwg-fetch": "^0.0.33",
-		"autoprefixer": "^10.4.16",
-		"daisyui": "^3.9.3",
-		"eslint": "^8",
-		"eslint-config-next": "13.5.4",
-		"eslint-plugin-jest-dom": "^5.0.2",
-		"eslint-plugin-testing-library": "^6.0.0",
-		"jest": "^29.7.0",
-		"jest-environment-jsdom": "^29.6.2",
-		"msw": "^1.3.2",
-		"next-router-mock": "^0.9.10",
-		"node-mocks-http": "^1.13.0",
-		"postcss": "^8.4.31",
-		"tailwindcss": "^3.3.5",
-		"ts-jest": "^29.1.1",
-		"ts-mockito": "^2.6.1",
-		"typescript": "^5",
-		"whatwg-fetch": "^3.6.19"
+		"@types/react-slick": "^0.23.13",
+		"postcss": "^8",
+		"tailwindcss": "^3.4.1",
+		"typescript": "^5.8.2"
 	}
 }
