@@ -1,22 +1,30 @@
-import type { Metadata } from 'next';
-import { Inter as FontSans } from 'next/font/google';
-import './globals.css';
-import { cn } from '@/lib/utils';
-import { ThemeProvider } from '@/components/theme-provider';
+import type { Metadata } from "next";
+import localFont from "next/font/local";
+import "./globals.css";
+import Header from "./components/Header/Header";
 
-import Link from 'next/link';
-import Header from '@/components/header';
-import Footer from '@/components/Footer';
-
-export const fontSans = FontSans({
-  subsets: ['latin'],
-  variable: '--font-sans',
+const poppins = localFont({
+  src: [
+    {
+      path: "../public/fonts/Recoleta-Black.ttf",
+      weight: "400",
+    },
+    {
+      path: "../public/fonts/Recoleta-Medium.ttf",
+      weight: "300",
+    },
+    {
+      path: "../public/fonts/Recoleta-Light.ttf",
+      weight: "200",
+    },
+  ],
+  variable: "--font-Recoleta-Black",
 });
 
 export const metadata: Metadata = {
-  title: 'Home | CKBEAM-SOFTWARE',
+  title: "Home | CKBEAM-SOFTWARE",
   description:
-    'Welcome to CKBEAM SOFTWARE, a personal portfolio website for Christian Beam.',
+    "Welcome to CKBEAM SOFTWARE, a personal portfolio website for Christian Beam.",
 };
 
 export default function RootLayout({
@@ -26,28 +34,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en" suppressHydrationWarning>
-      <body
-        className={cn(
-          'min-h-screen bg-background font-sans antialiased',
-          fontSans.variable
-        )}
-      >
-        <ThemeProvider
-          attribute="class"
-          defaultTheme="light"
-          enableSystem
-          disableTransitionOnChange
-        >
-          <header className="flex gap-4 justify-between items-center">
-            <Link className="ml-5 v-10" href="/">
-              CKBEAM-SOFTWARE
-            </Link>
-            <Header />
-          </header>
-          {children}
-          <Footer />
-        </ThemeProvider>
-      </body>
+      <body className={poppins.className}><Header/>{children}</body>
     </html>
   );
 }
