@@ -1,6 +1,6 @@
+// postcss.config.js
 module.exports = {
   plugins: {
     tailwindcss: {},
-    autoprefixer: {},
   },
-}
+};
